a little mod to 1

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -12,8 +12,8 @@
 
 char *_strncat(char *dest, char *src, int n)
 {
-	int i;
-	int l = strlen(src);
+	unsigned long i;
+	unsigned long l = strlen(src);
 	char c;
 	char *cadd;
